task: melhorias no core dos controllers
