+++ .gitignore
@@ -200,3 +200,6 @@ out/
 
 # mpeltonen/sbt-idea plugin
 .idea_modules/
+
+.idea
+*.mwb.bak
